chat channel user fixed
